Move custom call handler resolution into CustomCallThunk
For a custom call we need to resolve a target name (string) to a C++ function pointer. So far this happens in IrEmitterUnnested (at the end of the compilation pipeline). But for thunk serialization we need this to happen at runtime (when the thunks are getting reconstructed from the proto representation). Therefore I'm moving this resolving step into the CustomCallThunk factory function.

Note that there is still a way to construct a CustomCallThunk from just a function pointer. These thunks will not be serializable. I opted for not migrating those because all cases where we construct the thunk this way is in tests that use the legacy custom call API. These will need to get migrated to FFI in the next few months anyway and then this edge case goes away. Note that this is not a problem because we don't need the thunks in tests being serializable.

In detail this change entails:
- Move resolver logic in new overloads of the CustomCallThunk::Create factory function
- Call these overloads from IrEmitterUnnested and the custom kernel fusion emitter.
- Migrate some tests to FFI
- Adjust some error codes in custom call tests. (If a custom call was not found now kNotFound is returned instead of kUnimplemented).

PiperOrigin-RevId: 814223143